Make parameters for python mysql lib and node path configurable
When this role was run with python3 against an Ubuntu 20.04
machine, there were a couple of errors
where the code expected a different environment.
The role's python mysql package, python-mysqldb,
was specific to python 2, so would not be found
when ansible was run with python 3.
The role also looked for the node binary in /usr/bin/nodejs,
but the binary on ubuntu 20.04 is in /usr/bin/node.
This commit abstracts those parameters so that the user
can input values that match their own environment.
The parameters are now {{ etherpad_python_mysql_package }}
and {{ etherpad_node_path }} .